Sprint 3 User Stories: 2.5 (#45), 2 (#9), 1 (#5) Epic(s): 2 (#8), 1 (#3) Story Pts: 1
Summary:
I replaced the favicon logo with a higher quality version of the same
one. I added a small sub-heading prompting the user to select their
preferred quiz type with a color switch. I also added a GIF for flair.
This will hopefully aid in making the navigation more clear and
intuitive for the user.