Sprint 3 User Stories: 7 (#39), 6 (#34), 5 (#28) Epic(s): 5 (#33), 4 (#27) Story Pts: 4
Summary:
I added the result page, which you can reach by hitting the submit button on the basic/detailed questions. I integrated the ChatGPT API so it uses the answered saved in the local cache and send a request using the API. it returns 3 careers that best mataches the answers and give you a small description about it. I also added a button to "save" your result, with doesn't work but allows us to expend on it later.