Sprint 4 User Stories: 7 (#39), 2.5 (#45), 4 (#22), 1(#5) Epic(s): 1 (#3) Story Pts: 1
Summary:
I created this branch in order to finalize the site-wide adoption of the
navigation bar. Previously, every page on the site (About Us, Homepage,
Summary Pages, etc.) used the new navigation bar, except for the two
quizzes. I have now added them. I also added a LinkedIn and GitHub logo
to the About Us page next to the appropriate links. This is to enhance
the polish of the site, and uniformity for the user.